requirements.txt
PyQt5~=5.15.6
mysql-connector-python~=8.0.27
openpyxl~=3.0.9
requests~=2.26.0
fitz~=0.0.1.dev2
selenium~=4.1.0
keyboard~=0.13.5
future~=0.18.2
pyperclip~=1.8.2